chore(snap): code refactoring
